monitor/requirements.txt
Flask
requests
psycopg2-binary
mysql-connector-python
pytz
gunicorn